# docs/openapi-specification.yaml
openapi: 3.0.3
info:
  title: Ahamo Content Service API
  description: |
    ahamoダミーシステムのコンテンツサービスAPI仕様書
    
    このAPIは以下の機能を提供します：
    - キャンペーン情報の管理
    - ニュース情報の管理
    - FAQ情報の管理
    - お問い合わせ情報の管理
    
    新機能として、キャンペーンの有効性チェック機能を追加しました。
  version: 1.0.0
  contact:
    name: Ahamo Development Team
    email: dev@ahamo.com
servers:
  - url: http://localhost:8080
    description: ローカル開発環境
  - url: https://api.ahamo.com
    description: 本番環境

paths:
  /campaigns:
    get:
      summary: キャンペーン一覧取得
      description: ページネーション対応のキャンペーン一覧を取得します
      parameters:
        - name: page
          in: query
          description: ページ番号（0から開始）
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          description: 1ページあたりの件数
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: キャンペーン一覧取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignPageResponse'
        '400':
          description: リクエストパラメータエラー
        '500':
          description: サーバーエラー

  /campaigns/{id}:
    get:
      summary: キャンペーン詳細取得
      description: 指定されたIDのキャンペーン詳細情報を取得します
      parameters:
        - name: id
          in: path
          required: true
          description: キャンペーンID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: キャンペーン詳細取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignResponse'
        '404':
          description: キャンペーンが見つかりません
        '500':
          description: サーバーエラー

  /campaigns/{id}/validity:
    get:
      summary: キャンペーン有効性チェック
      description: |
        指定されたキャンペーンの有効性をチェックします。
        以下の条件を確認します：
        - キャンペーンがアクティブかどうか
        - 有効期間内かどうか（開始日・終了日）
      parameters:
        - name: id
          in: path
          required: true
          description: キャンペーンID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 有効性チェック結果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignValidityResponse'
        '400':
          description: 無効なキャンペーンID
        '500':
          description: サーバーエラー

  /news:
    get:
      summary: ニュース一覧取得
      description: ページネーション対応のニュース一覧を取得します
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: ニュース一覧取得成功

  /faqs:
    get:
      summary: FAQ一覧取得
      description: ページネーション対応のFAQ一覧を取得します
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: FAQ一覧取得成功

  /contacts:
    get:
      summary: お問い合わせ一覧取得
      description: ページネーション対応のお問い合わせ一覧を取得します
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: お問い合わせ一覧取得成功

  /health:
    get:
      summary: ヘルスチェック
      description: アプリケーションの稼働状況を確認します
      responses:
        '200':
          description: 正常稼働中

components:
  schemas:
    CampaignResponse:
      type: object
      description: キャンペーン情報レスポンス
      properties:
        id:
          type: string
          description: キャンペーンID
          example: "1"
        title:
          type: string
          description: キャンペーンタイトル
          example: "春の新規契約キャンペーン"
        description:
          type: string
          description: キャンペーン説明
          example: "新規契約で月額料金3ヶ月無料"
        imageUrl:
          type: string
          description: キャンペーン画像URL
          example: "https://example.com/campaign1.jpg"
        link:
          type: string
          description: キャンペーン詳細ページURL
          example: "https://ahamo.com/campaign/spring2024"
        createdAt:
          type: string
          format: date-time
          description: 作成日時
        updatedAt:
          type: string
          format: date-time
          description: 更新日時
        isActive:
          type: boolean
          description: アクティブフラグ
          example: true
        validFrom:
          type: string
          format: date-time
          description: 有効開始日時
          nullable: true
        validUntil:
          type: string
          format: date-time
          description: 有効終了日時
          nullable: true

    CampaignValidityResponse:
      type: object
      description: キャンペーン有効性チェック結果
      required:
        - campaignId
        - title
        - isValid
        - validityStatus
        - reason
      properties:
        campaignId:
          type: string
          description: キャンペーンID
          example: "1"
        title:
          type: string
          description: キャンペーンタイトル
          example: "春の新規契約キャンペーン"
        isValid:
          type: boolean
          description: キャンペーンが有効かどうか
          example: true
        validityStatus:
          type: string
          description: 有効性ステータス
          enum:
            - VALID
            - EXPIRED
            - NOT_STARTED
            - INACTIVE
            - NOT_FOUND
          example: "VALID"
        validFrom:
          type: string
          format: date-time
          description: 有効開始日時
          nullable: true
        validUntil:
          type: string
          format: date-time
          description: 有効終了日時
          nullable: true
        reason:
          type: string
          description: 有効性判定理由
          example: "キャンペーンは有効です"

    CampaignPageResponse:
      type: object
      description: キャンペーンページネーションレスポンス
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/CampaignResponse'
        page:
          type: integer
          description: 現在のページ番号
        size:
          type: integer
          description: ページサイズ
        totalElements:
          type: integer
          description: 総件数
        totalPages:
          type: integer
          description: 総ページ数

    ErrorResponse:
      type: object
      description: エラーレスポンス
      properties:
        error:
          type: string
          description: エラーコード
        message:
          type: string
          description: エラーメッセージ
        timestamp:
          type: string
          format: date-time
          description: エラー発生時刻

tags:
  - name: campaigns
    description: キャンペーン関連API
  - name: news
    description: ニュース関連API
  - name: faqs
    description: FAQ関連API
  - name: contacts
    description: お問い合わせ関連API
  - name: health
    description: ヘルスチェックAPI